requirements.txt
telethon
python-dotenv
gspread
google-auth
beautifulsoup4
pytz
requests

